requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
et_xmlfile==2.0.0
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.20.0
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
tomli==2.4.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
